[step2] refactor: 패키지 변경
