+++ src/main/kotlin/blackjack/card/SimpleCardCreator.kt
@@ -1,9 +1,4 @@
-package blackjack.factory
-
-import blackjack.card.Card
-import blackjack.card.CardSymbol
-import blackjack.card.PlayCards
-import blackjack.card.Suit
+package blackjack.card
 
 object SimpleCardCreator {
